test read token in server
